% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbID.R
\name{fbID}
\alias{fbID}
\title{Extract ID page facebook}
\usage{
fbID(x)
}
\arguments{
\item{x}{url facebook (\code{character}).}
}
\description{
Extract ID page facebook
}
\examples{
\dontrun{
fbID("https://www.facebook.com/denisson.silva.9")
}
}
