% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igID.R
\name{igID}
\alias{igID}
\title{Extract ID page Instagram}
\usage{
igID(x)
}
\arguments{
\item{x}{url instragram (\code{character}).}
}
\description{
Extract ID page Instagram
}
\examples{
\dontrun{
igID("https://www.instagram.com/denissonmatuto")
}
}
